Rag-Assistant/requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Embeddings & Vector Store
sentence-transformers==2.2.2
faiss-cpu==1.13.0

# UI (Optional)
streamlit==1.31.0

# Utilities
numpy==1.25.0

pandas==2.0.3
tqdm==4.66.1
